requirements.txt
flask
openai
mysql-connector-python